Fixed some directory stuff on Win32.
Fixed some directory stuff to do with the build on Win32.

+++ test_cpp/test_main.cpp
@@ -51,7 +51,7 @@ private:
     int three;
 
 public:
-    void inline_func(float one, float two)
+    void inline_func(int one, int two)
     {
         this->one = one;
         this->two = two;
